Web_app/requirements.txt
# Web Framework
fastapi==0.124.4
uvicorn==0.38.0
python-multipart==0.0.20
transformers==4.55.4
accelerate==1.12.0
torch==2.9.1+cu130
--extra-index-url https://download.pytorch.org/whl/cu130
optimum==2.0.0
peft==0.18.0
bitsandbytes==0.48.2
safetensors==0.7.0
llama-cpp-python==0.3.4
--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu124
sentencepiece==0.2.1
tiktoken==0.12.0
pandas==2.3.3
numpy==2.2.6
datasets==4.4.1
scikit-learn==1.7.2
matplotlib==3.10.7
seaborn==0.13.2
wordcloud==1.9.4
plotly==6.5.0
nltk==3.9.2
sentence-transformers==5.1.2
protobuf==5.29.5
tqdm==4.67.1
requests==2.32.5
python-dotenv==1.2.1
cmake==4.2.0